+++ lib/activesupport_cache_database.rb
@@ -1,1 +1,2 @@
 require 'active_support/cache/database_store'
+
